Stable version ready to merge

+++ src/components/Tasks/task.tsx
@@ -22,12 +22,16 @@ const Task: React.FC<TaskProps> = ({ task, setEditId }) => {
     }
 
     const handleDeleteTask = (e: React.FormEvent, id: string) => {
+        const opt = window.confirm("Please confirm you mean to delete this item?")
+        console.assert(opt)
+        if (!opt) {
+            return false
+        }
         e.preventDefault();
         const resp = DeleteTask(id)
         resp.then((result: ApiDataType | undefined) => {
             if (result) {
                 updateContexts(result.tasks)
-
             }
         })
             .catch((error) => {
@@ -45,17 +49,18 @@ const Task: React.FC<TaskProps> = ({ task, setEditId }) => {
             </div>
             <div className="Card--button">
                 <button
-                    onClick={(e) => setEditId(task._id)}
-                    className={"Card--button__done"}
-                >
-                    Edit
-                </button>
-                <button
                     onClick={(e) => handleUpdateTask(e, task)}
                     className={task.status ? `strike-button` : "Card--button__done"}
                 >
-                   {task.status ? `Uncomplete` : "Complete" }
+                    {task.status ? `Uncomplete` : "Complete"}
                 </button>
+                <button
+                    onClick={(e) => setEditId(task._id)}
+                    className={"Card--button__edit"}
+                >
+                    Edit
+                </button>
+
                 <button
                     onClick={(e) => handleDeleteTask(e, task._id)}
                     className="Card--button__delete"
